Fixed some methods of the inventory interface and added new test methods
(cherry picked from commit c7aac4a030bca560d929fa38103f41d108ec7f68)
